third_party/WebKit/LayoutTests/fast/dom/DOMException/dispatch-event-exception-expected.txt

Tests that dispatchEvent's argument is required to be an Event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.dispatchEvent(null) threw exception TypeError: Failed to execute 'dispatchEvent' on 'EventTarget': parameter 1 is not of type 'Event'..
PASS document.dispatchEvent(document) threw exception TypeError: Failed to execute 'dispatchEvent' on 'EventTarget': parameter 1 is not of type 'Event'..
PASS successfullyParsed is true

TEST COMPLETE
